requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2 
python-dotenv==1.0.1 
faiss-cpu==1.8.0 
langchain-community
langchain-groq 
langchain-huggingface 
langgraph==0.2.14 
transformers==4.44.2 
torch==2.4.1 
numpy==1.26.4 
sentence-transformers==3.0.1